1、修复docker元数据设置 2、修复jfrog pattern搜索 3、修复代理库收回站制品回滚 4、复制、移动性能优化

+++ folib-storage/folib-storage-api/src/main/java/com/veadan/folib/providers/io/StorageFileSystemProvider.java
@@ -168,10 +168,10 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath repositoryPath = (RepositoryPath) path;
-        if (!Files.exists(repositoryPath.getTarget())) {
-            throw new NoSuchFileException(unwrap(repositoryPath).toString());
-        }
-
+//        if (!Files.exists(repositoryPath.getTarget())) {
+//            throw new NoSuchFileException(unwrap(repositoryPath).toString());
+//
+//        }
         if (!Files.isDirectory(repositoryPath)) {
             doDeletePath(repositoryPath, force, true);
 
@@ -276,7 +276,9 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath trashPath = getTrashPath(repositoryPath);
-
+        if (!Files.exists(repositoryPath)) {
+            return;
+        }
         try {
             Files.move(repositoryPath.getTarget(),
                     trashPath.getTarget(),
